Fixed an off by one and mark searches as solved
The search had an off by one bug where the last character in the target string would never be evaluated.

This also caused searches to never be marked as solved.

Furthermore the state of the search will now be saved in the commissions to not start at zero once the search gets interrupted